refactor: Update user access for can_access field
This commit updates the user access functionality by adding support for the 'can_access' field. The 'changeUserAccess' function in the 'users.js' file now allows updating the 'can_access' value for a specific user identified by the UUID. The updated value is returned in the response along with a success toast message.