refactor: Update select routes to include material information
This commit updates the select routes in the `recipe_entry.js` file to include additional information about the material used in the recipe entry. The `material_uuid`, `material_name`, and `unit` fields are now included in the response. This change required importing the `materialSchema` module and adding a new join clause to fetch the material information.

Refactor the select routes to return the first row of data

This commit also refactors the select routes in the `recipe_entry.js` file to return only the first row of data. This improves performance by reducing the amount of data returned from the database.